player can add comment and rating only once after game

+++ src/main/java/sk/tuke/kpi/kp/service/RatingService.java
@@ -4,7 +4,7 @@ import sk.tuke.kpi.kp.entity.Rating;
 
 public interface RatingService {
     void setRating(Rating rating) throws RatingException;
-    int getAverageRating(String game) throws RatingException;
+    double getAverageRating(String game) throws RatingException;
     int getRating(String game, String player) throws RatingException;
     void reset() throws RatingException;
 }
